// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/driverSettlementSummaryResponse.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface DriverSettlementSummaryResponse {
  driverId?: number;
  totalSettlementCount?: number;
  pendingSettlementCount?: number;
  completedSettlementCount?: number;
  failedSettlementCount?: number;
  totalPayoutAmount?: number;
  pendingPayoutAmount?: number;
  completedPayoutAmount?: number;
  failedPayoutAmount?: number;
  monthPayoutAmount?: number;
  weekCompletedCount?: number;
  weekCompletedPayoutAmount?: number;
  calculatedAt?: string;
}